Separate NavigationTabBar depends on newDotHome beta
